+++ cpp/src/knn/knn_sparse.cu
@@ -27,7 +27,7 @@ namespace ML {
 namespace Sparse {
 
 void brute_force_knn(cumlHandle &handle, const int *idxIndptr,
-                     const int *idxIndices, const float *idxData, int idxNNZ,
+                     const int *idxIndices, const float *idxData, size_t idxNNZ,
                      int n_idx_rows, int n_idx_cols, const int *queryIndptr,
                      const int *queryIndices, const float *queryData,
                      size_t queryNNZ, int n_query_rows, int n_query_cols,
